Static constants for `Contact`
